Update script-AddEvidence.yml
Added a default timestamp of when the evidence is added to the board. I haven't found a good use case for manually entering a time for evidence to be found. Removed the 'occurred' arguments from the script.

Perhaps this could also just be added as a default option, while still allowing for an argument of 'occurred' to be entered.