Censys: (#4112)
* Censys:

1. fixed bug described in issue #18131:
In the Censys API, when there are no results for a search, it throws a 404. Demisto is interpreting this as a playbook error when it's actually just the regular behavior of Censys.

2. Beautified the code a bit (no logic changes)

* Censys:

1. fixed bug described in issue #18131:
In the Censys API, when there are no results for a search, it throws a 404. Demisto is interpreting this as a playbook error when it's actually just the regular behavior of Censys.

2. Beautified the code a bit (no logic changes)

* Censys bug (fixed - insecure was mandatory)

* Censys bug (fixed - insecure was mandatory)

* Censys - fix bug

* Censys - fix bug

* Censys - fix bug

* add subtype

* censys - bug fix, update changelog

* censys - CR changes

* Censys - bug fix

* Censys - bug fix

* Updated descriptions.

* Updated descriptions.