Remove newlines in markdown header
The Demisto entry resulting from a Splunk search may not be properly
formatted when the search string contains special characters like pipe,
backticks etc. Removing newlines from the markdown header resolves this
issue and allows for the resulting Entry/Note/Evidence to be converted
to HTML using the built-in mdToHtml regardless of the presence of special
characters in the query string.