Cloudshark integration (#4823)
* Initial commit of the CloudShark integration

Initial commit of the CloudShark integation for Demisto. This includes a
test playbook that runs through all of the commands for this integration
successfully when a new incident is created with an attached PCAP file.

* Add CloudShark Test Playbook to conf file

Add the test playbook for the CloudShark integration to the conf file.

* Add description for the CloudShark integration

Add a first draft of the description for the CloudShark integration.

* Add initial documentation template

Add the initial documentation template generated by Demisto using the
HTMLDocsAutomation command. This will still need to be updated and
filled out.

* Update test playbook configuration

Update the test playbook configuration to add a timeout and specify
nightly as true.

* Remove newlines from CloudShark description

Remove the newlines from the CloudShark description doc as these were
getting rendered in the markdown.

* Rename API Key to API Token

* Update README

Update the README file template filling in some initial data.

* Update test playbook to download test pcap

Update the test playbook for this integration so that it downloads a
test pcap from the Demisto repo on GitHub to use.

* Fix result of delete command

Had the wrong Demisto context being set. This should be the result of
the API call not just the Capture ID.

* Update output of upload command

Update the output of the upload command so that it shows a markdown link
to the CLOUDSHARK_URL configured and the capture id.

* Fix unclosed list

* Add list elements

* Update integration based on code review

* Added capture_id to human readable output of cloudshark-upload command
* Added check for API Token validity
* Added check for using bad URL
* Update to use handle_proxy()
* Changed CaptureId to CaptureID to meet output standard

* Update documentation based on internal review

Update the CloudShark integration documentatation based on our internal
review.

* Update documentation based on internal review

* Update based on CI tests failing

The CI tests were failing due to some code conventions that needed updating on our end.

* Remove whitespace from blank line

Remove whitespace from blank line

* Updates from Demisto code review

These are some additional updates from the code review.

* Print the reason of the error to the user
* Add outputs for sub object fields returned by cloudshark-info
* Change check_certificate to insecure
* Add test playbook

* Add Cloudshark-Test to Tests

Add the Cloudshark-Test to the Tests/conf.json file so that these tests are run.

* Add TestCloudshark to Tests

TestCloudshark was also missing in the Tests/conf.json file according to the CI results. This should add this.

* Fix playbookID of TestCloudshark playbook

Fix the playbook ID of the TestCloudshark playbook so that these tests run.

* Updated descriptions

Co-authored-by: Alex Fiedler <38628621+kirbles19@users.noreply.github.com>