[cofense-32] Two new commands and internal refactoring, second PR (#7104)
* [CofenseTriage] Add new Triage commands

on-behalf-of: @Cofense <oss@cofense.com>

* [CofenseTriage] WIP tests

on-behalf-of: @Cofense <oss@cofense.com>

* [CofenseTriage] Code style cleanup

on-behalf-of: @Cofense <oss@cofense.com>

* [CofenseTriage] assorted cleanup WIP

on-behalf-of: @Cofense <oss@cofense.com>

* [CofenseTriage] Add test fixtures WIP

on-behalf-of: @Cofense <oss@cofense.com>

* [CofenseTriage] Refactor http_request

- Rename to `triage_request` and rename first parameter to `endpoint`
- Create new function `triage_api_url` to build full URL to a given endpoint
- Refactor and simplify response handling logic

on-behalf-of: @Cofense <oss@cofense.com>

* [CofenseTriage] Refactor `fetch_reports`

No functional changes, except some speedups and a possible bug fix.

on-behalf-of: @Cofense <oss@cofense.com>

* [CofenseTriage] Break out TriageReport class

on-behalf-of: @Cofense <oss@cofense.com>

* [CofenseTriage] Break out TriageInstance class

on-behalf-of: @Cofense <oss@cofense.com>

* [CofenseTriage] Add TriageReporter class

on-behalf-of: @Cofense <oss@cofense.com>

* [CofenseTriage] Update tests and fixtures

on-behalf-of: @Cofense <oss@cofense.com>

* [CofenseTriage] Rewrite get_report_by_id to use class

on-behalf-of: @Cofense <oss@cofense.com>

* [CofenseTriage] Move all classes into one file

The plugin architecture requires it.

on-behalf-of: @Cofense <oss@cofense.com>

* [CofenseTriage] Refactor and add test coverage for get_threat_indicators()

on-behalf-of: @Cofense <oss@cofense.com>

* [CofenseTriage] Eleminate unnecessary get_attachment()

on-behalf-of: @Cofense <oss@cofense.com>

* [CofenseTriage] Refactor search_reports and increase test coverage

on-behalf-of: @Cofense <oss@cofense.com>

* [CofenseTriage] Assume /reports/:id always returns an array

Also eliminate unnecessary TriageReporter.from_json() and rename
Triage_reporter.from_id() to .fetch().

on-behalf-of: @Cofense <oss@cofense.com>

* [CofenseTriage] Make test fixture more complicated

on-behalf-of: @Cofense <oss@cofense.com>

* [CofenseTriage] Return actual JSON in to_json()

Also enhance test coverage.

on-behalf-of: @Cofense <oss@cofense.com>

* [CofenseTriage] Don't call fetch_reports() in test_function()

Triage always responds with a valid JSON object. There is no need to
perform a second request to test the integration---if Triage responsds
with an OK status, then everything is working.

on-behalf-of: @Cofense <oss@cofense.com>

* [CofenseTriage] Store last run data as a JSON blob

on-behalf-of: @Cofense <oss@cofense.com>

* [CofenseTriage] Wrap incident attachment in single-element list

on-behalf-of: @Cofense <oss@cofense.com>

* [CofenseTriage] Update metadata

on-behalf-of: @Cofense <oss@cofense.com>

* [CofenseTriage] Clean up remnants in Legacy pack

on-behalf-of: @Cofense <oss@cofense.com>

* [CofenseTriage] Add mypy ignore comments

Mypy has trouble with decorators like lru_cache() in several situations.
Add inline comments to silence spurious linter complaints.

on-behalf-of: @Cofense <oss@cofense.com>

* [CofenseTriage] Move import after stubs in test

We have to stub demistomock before we import CofenseTriage. That's just how
demistomock works, apparently.

on-behalf-of: @Cofense <oss@cofense.com>

* [CofenseTriage] Generate release notes

on-behalf-of: @Cofense <oss@cofense.com>

* [CofenseTriage] Move changes to new CofenseTriage 2

Both versions will exist in parallel

on-behalf-of: @Cofense <oss@cofense.com>

* [CofenseTriage] Move return_error to highest-level except block

on-behalf-of: @Cofense <oss@cofense.com>

* [CofenseTriage] Update documentation

on-behalf-of: @Cofense <oss@cofense.com>

* [CofenseTriage] Address various linter complaints

on-behalf-of: @Cofense <oss@cofense.com>

* [CofenseTriage] Move time constants inline

Also eliminate the time format string in favor of
datetime.datetime.fromisoformat().

on-behalf-of: @Cofense <oss@cofense.com>

* [CofenseTriage] Pass a TriageInstance argument instead of using a module var

on-behalf-of: @Cofense <oss@cofense.com>

* [CofenseTriage] Let exceptions bubble up to main()

on-behalf-of: @Cofense <oss@cofense.com>

* [CofenseTriage] Address more linter complaints

on-behalf-of: @Cofense <oss@cofense.com>

* [CofenseTriage] Rename module to CofenseTriagev2

on-behalf-of: @Cofense <oss@cofense.com>

* [CofenseTriage] Centralize parameter fetching in main()

on-behalf-of: @Cofense <oss@cofense.com>

* [CofenseTriage] Add v2 to Tests/conf.json

on-behalf-of: @Cofense <oss@cofense.com>

* [CofenseTriage] Rename more files to have v2 prefix

on-behalf-of: @Cofense <oss@cofense.com>

* [CofenseTriage] Add minimum Demisto version

on-behalf-of: @Cofense <oss@cofense.com>

* [CofenseTriage] Move test files to root dir of integration

on-behalf-of: @Cofense <oss@cofense.com>

* [CofenseTriage] Remove tests from v1 integration

on-behalf-of: @Cofense <oss@cofense.com>

* [CofenseTriage] Rename Cofense.ThreatIndicators context path

on-behalf-of: @Cofense <oss@cofense.com>

* [CofenseTriage] Additional minor adjustments

on-behalf-of: @Cofense <oss@cofense.com>

* [CofenseTriage] Merge all test files into one

on-behalf-of: @Cofense <oss@cofense.com>

* [CofenseTriage] Stub fileResult more realistically

on-behalf-of: @Cofense <oss@cofense.com>

* [CofenseTriage] Update release notes

on-behalf-of: @Cofense <oss@cofense.com>

Co-authored-by: Mike Saurbaugh <mike.saurbaugh@cofense.com>