Update cisco-ise.py (#9601)
* Update cisco-ise.py

Added two api's one ERS api (/ers/config/ancendpoint/clear) and one admin/monitoring api (/admin/API/mnt/Session/EndPointIPAddress/{ip}). The "clear" api removes a policy that has been assigned to an endpoint, the "session" api returns the session data from an active endpoint and writes this data to the context menu to be used as data enrichment or to access data from the session as needed. Because the admin/monitoring api's don't use the port (9060 for ERS api calls) a new url (without the port assigned) was created, a new http_request function was added to handle the admin calls and a new header to handle returned xml was added.

* updated yml file, added release notes

* reverted optional argument

Co-authored-by: syaakovi <syaakovi@paloaltonetworks.com>