Main Changes In This PR: - Env variables - before_script - cache - stages
Env Variables:
- `DONT_CACHE_LAST_RESPONSE` - Handles the memory leak in the demisto-py client.
- `ARTIFACTS_FOLDER` - The new path of the build's artifacts.
- `BASH_ENV` - A file that will contain dynamic env variables that will be passed on between jobs via artifacts.
- `GIT_STRATEGY` and `GIT_DEPTH` - Gitlab has some optimizations for builds that needs only the files of the build's branch. This is not the case for us as we need a full copy in order to to make our diff calls. These variables tells gitlab to use a full clone on the initialization.

Cache:
Should work pretty much the same as in circle, but the cache is saved locally on the gitlab runners, this will help optimising as we will tun out builds on more and more runners and as we have workflows with more than one job.

Image:
 - The `devdemisto/gitlab-content-ci:1.0.0.18863` was created by @DinaMeylakh in [this](https://github.com/demisto/dockerfiles/pull/4873) PR.
- The `docker-io.artifactory.pan.local` prefix makes the image pull be authenticated and overcome the docker rate limit.

before_script:
The `before_script` is a gitlab feature and it contains steps that will be executed before the job's steps.
It prepares the environment pretty much like in circle jobs with the following exceptions:
- The `content-test-conf` repo is downloaded in here and not in the job's steps.
- The ssh keys are loaded in here.