SDP Create Request, fix POST request
The current POST query puts input_data in the URL parameters. Although this is supported, it may create errors mentionning 'URI too long' if input data is too long. This commit introduces the use of the POST request body instead of the URL parameters to pass the input_data
corrected the import of json