Update README.md (#15274)
Added the following to the Listen Port description in light of https://github.com/demisto/etc/issues/41066:

You can use any available port except for 80, 443, or 9100. 
When the `instance.execute.external.<instance_name>` key is set to true, Cortex XSOAR redirects the endpoint from HTTPS to the container on the port that you specify here, using port 443 as the secured publicly open port.