Clarification in step 1
Small suggestion based on my first time following these instructions: It would minimize confusion to clarify that the user only needs to follow the linked Microsoft article through the Register an application section (https://docs.microsoft.com/en-us/azure/active-directory/develop/quickstart-register-app). The steps beyond that are either unnecessary or, in the case of adding a client secret, described in the next step in the XSOAR doc.