Aburt demisto panorama update - polling commands (#15063)
* Added polling functionality to commands that require polling for
the results. Also added generic input for the "panorama" command to
support generic input and also an entryID input to support files in
the requests library command.

* Added polling functionality to commands that require polling for
the results. Also added generic input for the "panorama" command to
support generic input and also an entryID input to support files in
the requests library command.

* Added polling functionality to commands that require polling for
the results. Also added generic input for the "panorama" command to
support generic input and also an entryID input to support files in
the requests library command.

* Added polling functionality to commands that require polling for
the results. Also added generic input for the "panorama" command to
support generic input and also an entryID input to support files in
the requests library command.

* Update Packs/PAN-OS/Integrations/Panorama/Panorama.py

* Update Packs/PAN-OS/Integrations/Panorama/Panorama.py

* Added the run_polling_command command to handle all polling requests
from a single function.

* Removed unused result_path in panorama_query_logs_command

* Apply suggestions from code review

Co-authored-by: Bar Katzir <37335599+bakatzir@users.noreply.github.com>

* Apply suggestions from code review

Co-authored-by: Bar Katzir <37335599+bakatzir@users.noreply.github.com>

* Apply suggestions from code review

Co-authored-by: Dean Arbel <darbel@paloaltonetworks.com>

* CHanged inline dictionary iteration to built-in function of
dict_safe_get from CommonServerPython

* Update Packs/PAN-OS/Integrations/Panorama/Panorama.py

Co-authored-by: Adam Burt - Demisto <aburt-demisto@users.noreply.github.com>
Co-authored-by: Bar Katzir <37335599+bakatzir@users.noreply.github.com>
Co-authored-by: Dean Arbel <darbel@paloaltonetworks.com>