Update README.md
Added missing descriptions. Updated existing descriptions with more detail. Added some clarity to a Note that used the term "we", which is confusing. Added additional topic, which displays in Help tab but is best here--in one location. I still don't know what to write for Use system proxy settings, but will figure it out and update with next pull request.