Fixed misleading arg description (#18063)
Fixed misleading arg description for `microsoft-teams-member-add` arg `user_id`. We were initially thrown off because the arg description says the value needs to be an ID that can only be retrieved by running a different command where the user is already a member of a team. However, we tested passing the user's email address as the `user_id` and this worked. Perhaps this option was added later on, and just the arg description was not updated. Not sure if the same thing applies to other commands in this integration.