Update README.md (#18458)
Is there an appropriate section of this document to add this information:

EDL is designed to spawn on two processes: nginx and python. Nginx is the process that listens on the configured port, while the python process listens on the configured port + 1. This means that if an integration was configured for port 9009, the nginx process will listen on port 9009 and python on port 9010.
When running without --network=host the python port is not exposed to the machine.