netconn local ip remote ip fix
There are two different formats for netconn_complete. The v1 API returns an array of pipe delimited strings:

"2013-09-16 07:11:59.000000|-1979811809|80|6|dl.javafx.com|true"

    field 0: event time
    field 1: remote IP address as a 32-bit signed long (host byte order)
    field 2: remote port
    field 3: protocol: 6 is TCP, 17 is UDP
    field 4: domain name associated with the IP address, from the client’s perspective at the time of the network connection
    field 5: boolean “true” if the connection was outbound; “false” if the connection was inbound

The v2 API and newer return an array of JSON objects:

[{"domain": "login.live.com",
  "proto": 6,
  "local_port": 49240,
  "timestamp": "2017-01-11T16:20:04.892Z",
  "local_ip": 167772448,
  "direction": "true",
  "remote_port": 80,
  "remote_ip": -2080555708},
  ...
]

Since the request is made to v3 there is no parser implemented for netconn_complete as it's not needed. The caveat to this however is that IPs are in integer format.