Update README.md
Added link to MS document that a client passed along, which allowed them to limit the application's access to specific mailboxes only. They confirmed that this worked to limit the application's access. They had reached out to ask about this because it was not clear from the docs whether this was possible.