Add timeout functionality to rasterize_email_command (#18647)
Currently the rasterize_email_command() function does not have any functionality to alter the timeout from the default of 180 set in the rasterize() function. This can be fixed by using the same logic for setting the max_page_load_time found in the rasterize_command() function. This change will help allow clients to alter the timeout for HTML rasterization of larger emails that sometimes fall outside of the 180 second default.

Another option for this fix would be adding an argument to the function within the integration to allow this variable to be set dynamically instead of using the default set in the integration params (which is what this submitted request would end up defaulting to).