Update SendEmailReply.py
These statements were nested in the previous conditional if statement "if attachments". This led to an issue where email sent by the user in response to an email would not be properly tagged unless an attachment was included in the response. Moving these statements out from this condition will correct this error.