More defensive handling of NoneType values
Adds defense-in-depth for dealing with potential NoneType values

If a required key is missing or has a NoneType value, it will be overwritten either by: (1) an empty string if it is the only value for that owner or (2) another value via aggregation. In case (1) we are unfortunately mutating the data from None to '', but I think this case is rare and unlikely to cause downstream trouble; preserving the None seems like more trouble than it's worth.

Test plan: pytest