Allow optional fallback for body extraction.
If the input html has no body tag, but represent the body
(e.g. the content of an email), this commit add the option,
to use the full html as a fallback if no body tag was found.

The added argument is optional and defaults to 'false' to
ensure backwards compability.

The fallback only applies, if html_tag is 'body'