[TextFromHTML] Rundssoar contrib automations (#29836)
* use regex flags for html-tag replacement as well.
HTML from unknown sources, e.g. emails, may contain upper-case html tags

* sanitize whitespaces in converted text.
To keep the output more readabe:
- replace multible successive space and tabs with a single space
- replace multible successive linebreaks with a single linebreak

* Allow optional fallback for body extraction.

If the input html has no body tag, but represent the body
(e.g. the content of an email), this commit add the option,
to use the full html as a fallback if no body tag was found.

The added argument is optional and defaults to 'false' to
ensure backwards compability.

The fallback only applies, if html_tag is 'body'

* Allow to output the extracted Text to the Contex
The new argument is optional to ensure backwards compability

* Relax newline sanitization
Allow double new line.
Collaps more empty lines to double empty lines

* fix linting warnings

* Improve text extraction
- properly convert br tags to line breaks
- remove trailing whitespaces from end of extracted text

* Update README documentation

* update container to demisto/python3:3.10.13.74666

* Update and add Tests
- Test behavior for allow_body_fallback argument
- Update Tests for change with removed trailing whitespaces
- Add Test for properly handling of 'br' tags

* revert "save" access to html argument

We want it to fail if it doesn't exist for some reason (and not assign it an empty string value)/

Co-authored-by: Michael Yochpaz <8832013+MichaelYochpaz@users.noreply.github.com>

* Ignore unicode character check for unittest

ignore this rule for the whole file, since there are multiple tests testing non-unicode characters.

Co-authored-by: Michael Yochpaz <8832013+MichaelYochpaz@users.noreply.github.com>

* remove uncessary single line ignore

The error has been ignored for the whole file, so no need to have it as a lingle line ignore

Co-authored-by: Michael Yochpaz <8832013+MichaelYochpaz@users.noreply.github.com>

* make linebreak replacement and trimming optional
- linebreaks can optionally be replaced via argument
- trimming of result text and consecutive whitespaces is optional
- updated br-tag regex to match all valid variants
- updated mutlible linebreak regex to consider lines with only
whitespace as an empty line

* reduce number of Regex Flags
the IgnoreCase Flag remains, because input HTML may contain both uppercase
and lowercase characters that are aren't handled elsewhere and may lead
to incomplete removal of tags

* Cleanup main function and use fix context path
Outputting to the context remains optional, so existing playbooks
are not changed.
This also adds the proper output to the command.

* Fix: use proper data variable for tag removal

* update readme

* Fix broken tests after introduction of main
in 5074e5c, the text_from_html function
was merged into the new main function.
This broke the unit tests. This commit updates
the unit tests to work with the new functions.

* Add test for trimmed result and make function parameter
explicit to improve readability

* fix pre-commit check fail in yaml

* fix yaml formating

* Revert "fix yaml formating"

This reverts commit 26a31d2ba29b8b16cd8998f903af43c869422683.

* remove unnecessary Regex flag

* Fix unittests after making trimming optional

* fix unit missed in last commit

* use argToBoolean instead of string compare

Co-authored-by: Michael Yochpaz <8832013+MichaelYochpaz@users.noreply.github.com>

* Apply suggestions from yaml review

Co-authored-by: Michael Yochpaz <8832013+MichaelYochpaz@users.noreply.github.com>

* add unittests for get_body

* exclude main from unit-tests coverage

The main function is not relevant for unittest

Co-authored-by: Michael Yochpaz <8832013+MichaelYochpaz@users.noreply.github.com>

* update readme

* update release notes

* apply format

* fix typo in pragma to skip testing for main

* update wording in releasenotes

Co-authored-by: ShirleyDenkberg <62508050+ShirleyDenkberg@users.noreply.github.com>

* update Readme and yaml

update the wording for descriptions of arguments in the readme and the yaml file
Also remove testing playbook that is not relevant for users in the readme

Co-authored-by: Michael Yochpaz <8832013+MichaelYochpaz@users.noreply.github.com>
Co-authored-by: ShirleyDenkberg <62508050+ShirleyDenkberg@users.noreply.github.com>

* update docker image to 3.10.13.75921

* update relase notes regarding docker image

---------

Co-authored-by: Michael Yochpaz <8832013+MichaelYochpaz@users.noreply.github.com>
Co-authored-by: ShirleyDenkberg <62508050+ShirleyDenkberg@users.noreply.github.com>