Anomali ThreatStream change DBot verdict from Benign to Unknown for Low Confidence Indicators (#30993)
* change DBot verdict from Benign to Unknown for Low Confidence Indicators

Indicators found in Anomali that are below Confidence thresholds should be created as Unknown and not Benign.  

Anomali ThreatStream documentation regarding Confidence
https://ui.threatstream.com/optic-doc/Content/Features/threat_model/Observables/details_indicator.htm

Confidence - Confidence indicates the certainty that an observable exhibits or is connected to malicious behavior.

If Anomali has indicators with low Confidence, that doesn't mean the indicator is Benign/Safe.  It means Anomali is unsure that the indicator is Malicious and as such the more appropriate verdict in XSOAR should be Unknown.

* add indicator_default_score param

* changed values to Benign and Unknown

* update README and RN

* update RN

* update docker
set required to false

* update docker

* fix docs comments

---------

Co-authored-by: adi88d <adaud@paloaltonetworks.com>