Update Brandefense.py
Hello I have updated the test module when an API key entered incorrectly 401/403 errors it would pop-up an authorization error string if anything else occurs rather than 401/403,200 it will pop-up the error with the status code in the test_module function. I think the PR is ready to merge at this point. :)